add avg layer size

+++ consensus/ninja_tortoise_test.go
@@ -93,7 +93,7 @@ func TestNinjaTortoise_UpdatePatternTally(t *testing.T) {
 func TestNinjaTortoise_Sanity1(t *testing.T) {
 	layerSize := 30
 	patternSize := layerSize
-	alg := NewNinjaTortoise(uint32(layerSize), log.New("TestNinjaTortoise_Sanity1", "", ""))
+	alg := NewNinjaTortoise(layerSize, log.New("TestNinjaTortoise_Sanity1", "", ""))
 	l1 := GenesisLayer()
 	genesisId := l1.Blocks()[0].ID()
 	alg.handleIncomingLayer(l1)
@@ -116,7 +116,7 @@ func TestNinjaTortoise_Sanity1(t *testing.T) {
 //vote explicitly for two previous layers
 //correction vectors compensate for double count
 func TestNinjaTortoise_Sanity2(t *testing.T) {
-	alg := NewNinjaTortoise(uint32(3), log.New("TestNinjaTortoise_Sanity2", "", ""))
+	alg := NewNinjaTortoise(3, log.New("TestNinjaTortoise_Sanity2", "", ""))
 	l := createMulExplicitLayer(0, map[mesh.LayerID]*mesh.Layer{}, nil, 1)
 	l1 := createMulExplicitLayer(1, map[mesh.LayerID]*mesh.Layer{l.Index(): l}, map[mesh.LayerID][]int{0: {0}}, 3)
 	l2 := createMulExplicitLayer(2, map[mesh.LayerID]*mesh.Layer{l1.Index(): l1}, map[mesh.LayerID][]int{1: {0, 1, 2}}, 3)
